Fixed bug in facebook login

+++ src/dealstealunreal.com/Models/User/Register.cs
@@ -54,6 +54,6 @@
         /// <summary>
         /// Facebook ID
         /// </summary>
-        public int FacebookId { get; set; }
+        public long FacebookId { get; set; }
     }
 }
